Added conditional check to ignore .ini files, since the hidden windows file 'desktop.ini' was causing the script to error out;  Changed project name in package-lock.json.

+++ upscale_backlog.js
@@ -47,37 +47,40 @@ const prep_files = () => {
 
 	// Run through all files in the folder
 	file_names.forEach((filename) => {
-		// Change new image type to .png
-		let last_dot = filename.lastIndexOf('.');
-		let png_file = filename.substr(0, last_dot) + '.png';
-		
-		// Get old and new filepaths
-		let file_path = folder_to_upscale + "\\" + filename;
-		let new_path = save_folder + "\\" + png_file;
-
-		// Get height and width of image
-		let dimensions = sizeOf(file_path);
-		let height = dimensions.height;
-		let width = dimensions.width;
-
-		// If width or height already >= 4K, just move the file to the save_folder
-		if (height >= 2160 || width >= 3840) {
-			// Copy the file to save_folder
-			fs.copyFileSync(file_path, new_path);
+		// Ignore annoying desktop.ini file, just ignore all .ini I guess
+		if (!filename.includes('.ini')) {
+			// Change new image type to .png
+			let last_dot = filename.lastIndexOf('.');
+			let png_file = filename.substr(0, last_dot) + '.png';
 			
-			// Delete the file from folder_to_upscale
-			fs.unlinkSync(file_path);
-		}
-		else {
-			// Find x/y scale to make image 4K in one direction
-			// Use the lower of the two values
-			let height_scale = Math.ceil(10 * (2160 / height)) / 10;
-			let width_scale = Math.ceil(10 * (3840 / width)) / 10;
+			// Get old and new filepaths
+			let file_path = folder_to_upscale + "\\" + filename;
+			let new_path = save_folder + "\\" + png_file;
 
-			let img_scale = height_scale < width_scale ? height_scale : width_scale;
+			// Get height and width of image
+			let dimensions = sizeOf(file_path);
+			let height = dimensions.height;
+			let width = dimensions.width;
 
-			// Add file paths/img_scale object to array
-			files_to_upscale.push({ img_scale: img_scale, file_path: file_path, new_path: new_path });
+			// If width or height already >= 4K, just move the file to the save_folder
+			if (height >= 2160 || width >= 3840) {
+				// Copy the file to save_folder
+				fs.copyFileSync(file_path, new_path);
+				
+				// Delete the file from folder_to_upscale
+				fs.unlinkSync(file_path);
+			}
+			else {
+				// Find x/y scale to make image 4K in one direction
+				// Use the lower of the two values
+				let height_scale = Math.ceil(10 * (2160 / height)) / 10;
+				let width_scale = Math.ceil(10 * (3840 / width)) / 10;
+
+				let img_scale = height_scale < width_scale ? height_scale : width_scale;
+
+				// Add file paths/img_scale object to array
+				files_to_upscale.push({ img_scale: img_scale, file_path: file_path, new_path: new_path });
+			}
 		}
 	});
